handle device config changes

+++ libplatune/player/src/channel_buffer.rs
@@ -1,5 +1,5 @@
 pub(crate) struct ChannelBuffer {
-    inner: Vec<Vec<f64>>,
+    inner: Vec<Vec<f32>>,
     capacity: usize,
     channels: usize,
     current_chan: usize,
@@ -19,11 +19,11 @@ impl ChannelBuffer {
         self.inner[self.channels - 1].len()
     }
 
-    fn last(&self) -> &Vec<f64> {
+    fn last(&self) -> &Vec<f32> {
         self.inner.last().expect("Vec should not be empty")
     }
 
-    fn next_chan(&mut self) -> &mut Vec<f64> {
+    fn next_chan(&mut self) -> &mut Vec<f32> {
         let chan = &mut self.inner[self.current_chan];
         self.current_chan = (self.current_chan + 1) % self.channels;
         chan
@@ -44,7 +44,7 @@ impl ChannelBuffer {
         self.current_chan = 0;
     }
 
-    pub(crate) fn inner(&self) -> &[Vec<f64>] {
+    pub(crate) fn inner(&self) -> &[Vec<f32>] {
         &self.inner
     }
 
@@ -54,7 +54,7 @@ impl ChannelBuffer {
         }
     }
 
-    pub(crate) fn fill_from_slice(&mut self, data: &[f64]) -> usize {
+    pub(crate) fn fill_from_slice(&mut self, data: &[f32]) -> usize {
         let mut i = 0;
         while self.len() < self.capacity && i < data.len() {
             self.next_chan().push(data[i]);
