+++ libplatune/player/src/dto/queue_source.rs
@@ -3,7 +3,7 @@ use crate::{settings::Settings, source::Source};
 #[derive(Debug)]
 pub(crate) struct QueueSource {
     pub(crate) source: Box<dyn Source>,
-    pub(crate) volume: Option<f64>,
+    pub(crate) volume: Option<f32>,
     pub(crate) settings: Settings,
     pub(crate) queue_start_mode: QueueStartMode,
 }
