+++ libplatune/player/src/vec_ext.rs
@@ -1,9 +1,9 @@
 pub(crate) trait VecExt {
-    fn fill_from_deinterleaved(&mut self, channels: &[Vec<f64>]);
+    fn fill_from_deinterleaved(&mut self, channels: &[Vec<f32>]);
 }
 
-impl VecExt for Vec<f64> {
-    fn fill_from_deinterleaved(&mut self, deinterleaved: &[Vec<f64>]) {
+impl VecExt for Vec<f32> {
+    fn fill_from_deinterleaved(&mut self, deinterleaved: &[Vec<f32>]) {
         let out_len = deinterleaved[0].len();
         self.clear();
